# facturation-planning-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour gérer la facturation et les plannings",
        "title": "Facturation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/clients/{id}/devis": {
            "get": {
                "description": "Récupère la liste de tous les devis appartenant à un client spécifique",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Récupérer tous les devis d'un client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du client",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des devis du client",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Devis"
                            }
                        }
                    },
                    "400": {
                        "description": "ID client manquant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/devis": {
            "get": {
                "description": "Récupère la liste complète de tous les devis avec leurs lignes, entreprises et clients",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Récupérer tous les devis",
                "responses": {
                    "200": {
                        "description": "Liste de tous les devis avec totaux calculés",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Devis"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des devis",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Crée un nouveau devis avec validation des données d'entreprise et de client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Créer un nouveau devis",
                "parameters": [
                    {
                        "description": "Données du devis à créer",
                        "name": "devis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Devis créé avec succès",
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    },
                    "400": {
                        "description": "Erreur de validation des données",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/devis/{id}": {
            "get": {
                "description": "Récupère un devis spécifique avec ses lignes, entreprise et client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Récupérer un devis par son ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis à récupérer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devis trouvé avec totaux calculés",
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    },
                    "404": {
                        "description": "Devis introuvable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Met à jour un devis existant avec validation des données",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Mettre à jour un devis existant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis à modifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles données du devis",
                        "name": "devis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devis mis à jour avec succès",
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    },
                    "400": {
                        "description": "Erreur de validation des données",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la mise à jour",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Supprime définitivement un devis du système",
                "tags": [
                    "Devis"
                ],
                "summary": "Supprimer un devis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis à supprimer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Devis supprimé avec succès"
                    },
                    "500": {
                        "description": "Erreur lors de la suppression",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/devis/{id}/download": {
            "get": {
                "description": "Génère et force le téléchargement d'un devis au format PDF",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Télécharger un devis en PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis à télécharger",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF à télécharger",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Devis introuvable",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération du PDF",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/devis/{id}/pdf": {
            "get": {
                "description": "Génère un devis au format PDF pour affichage dans le navigateur",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Générer un PDF de devis pour affichage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis à convertir en PDF",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF généré pour affichage",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Devis introuvable",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la génération du PDF",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/devis/{id}/statut": {
            "patch": {
                "description": "Met à jour uniquement le statut d'un devis (brouillon, envoyé, accepté, refusé, expiré)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devis"
                ],
                "summary": "Mettre à jour le statut d'un devis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID du devis",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouveau statut du devis",
                        "name": "statut",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statut": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devis avec statut mis à jour",
                        "schema": {
                            "$ref": "#/definitions/models.Devis"
                        }
                    },
                    "400": {
                        "description": "Statut invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la mise à jour du statut",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Devis": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/models.Client"
                },
                "client_id": {
                    "type": "integer"
                },
                "conditions": {
                    "type": "string",
                    "example": "Paiement sous 30 jours"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                },
                "date_devis": {
                    "type": "string",
                    "example": "2025-06-10T00:00:00Z"
                },
                "date_expiration": {
                    "type": "string",
                    "example": "2025-07-10T00:00:00Z"
                },
                "entreprise": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Entreprise"
                        }
                    ]
                },
                "entreprise_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "Champs GORM",
                    "type": "integer",
                    "example": 1
                },
                "lignes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LigneDevis"
                    }
                },
                "objet": {
                    "type": "string",
                    "example": "Développement application web"
                },
                "sous_total_ht": {
                    "description": "Champs calculés (ne pas stocker en base)",
                    "type": "number"
                },
                "statut": {
                    "description": "brouillon, envoyé, accepté, refusé",
                    "type": "string",
                    "example": "brouillon"
                },
                "total_ttc": {
                    "type": "number"
                },
                "total_tva": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                }
            }
        },
...
    }
}